+++ src/test/java/commonGraphics/panels/sidePanels/SectionContainerInterface.java
@@ -1,0 +1,6 @@
+package commonGraphics.panels.sidePanels;
+
+//
+public interface SectionContainerInterface {
+    void addSections();
+}
