--- src/test/java/planetTest/graphics/panels/leftSidePanel/SectionContainer.java
@@ -1,6 +1,0 @@
-package planetTest.graphics.panels.leftSidePanel;
-
-//
-public interface SectionContainer {
-    void addSections();
-}
