+++ src/test/java/planetTest/graphics/panels/leftSidePanel/ThirdSectionPanel.java
@@ -7,6 +7,7 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import static commonGraphics.StringUtils.drawNumberedString;
+import commonGraphics.panels.sidePanels.AbstractSectionPanel;
 
 //a blank section for future use
 final class ThirdSectionPanel extends AbstractSectionPanel {
